+++ petastorm/reader.py
@@ -32,7 +32,6 @@ from petastorm.predicates import PredicateBase
 from petastorm.py_dict_reader_worker import PyDictReaderWorker
 from petastorm.reader_impl.arrow_table_serializer import ArrowTableSerializer
 from petastorm.reader_impl.pickle_serializer import PickleSerializer
-from petastorm.reader_impl.pyarrow_serializer import PyArrowSerializer
 from petastorm.selectors import RowGroupSelectorBase
 from petastorm.transform import transform_schema
 from petastorm.workers_pool.dummy_pool import DummyPool
@@ -88,8 +87,7 @@ def make_reader(dataset_url,
         denoting a thread pool, process pool, or running everything in the master thread. Defaults to 'thread'
     :param workers_count: An int for the number of workers to use in the reader pool. This only is used for the
         thread or process pool. Defaults to 10
-    :param pyarrow_serialize: Whether to use pyarrow for serialization. Currently only applicable to process pool.
-        Defaults to False.
+    :param pyarrow_serialize: THE ARGUMENT IS DEPRECATED AND WILL BE REMOVED IN FUTURE VERSIONS.
     :param results_queue_size: Size of the results queue to store prefetched row-groups. Currently only applicable to
         thread reader pool type.
     :param shuffle_row_groups: Whether to shuffle row groups (the order in which full row groups are read)
@@ -152,9 +150,9 @@ def make_reader(dataset_url,
         reader_pool = ThreadPool(workers_count, results_queue_size)
     elif reader_pool_type == 'process':
         if pyarrow_serialize:
-            serializer = PyArrowSerializer()
-        else:
-            serializer = PickleSerializer()
+            warnings.warn("pyarrow_serializer was deprecated and will be removed in future versions. "
+                          "The argument no longer has any effect.")
+        serializer = PickleSerializer()
         reader_pool = ProcessPool(workers_count, serializer, zmq_copy_buffers=zmq_copy_buffers)
     elif reader_pool_type == 'dummy':
         reader_pool = DummyPool()